Update PDF tools extension: bump dependencies, refactor utility functions, and implement Windows-specific PDF handling
- Bumped @libpdf/core to version 0.2.9 and eslint to version 9.39.3 in package.json.
- Removed utils.ts and moved Windows-specific PDF splitting functions to a new windows.ts file.
- Updated imports in merge.ts, protect.ts, split-by-file-size.ts, split-by-page-count.ts, unlock.ts, and watermark.ts to use constants from the new constants.ts file.
- Added constants.ts to define platform checks for Mac and Windows.
