+++ extensions/pdf-tools/src/lib/constants.ts
@@ -1,0 +1,2 @@
+export const isMac = process.platform === "darwin";
+export const isWindows = process.platform === "win32";
