+++ extensions/pdf-tools/src/merge.ts
@@ -1,9 +1,8 @@
 import { closeMainWindow, getSelectedFinderItems, LaunchProps, showToast, Toast } from "@raycast/api";
 import fs from "fs/promises";
 import path from "path";
-import { PDF } from "@libpdf/core";
 import { getSelectedItems } from "universal-selection";
-import { isMac, isWindows } from "./lib/utils";
+import { isMac, isWindows } from "./lib/constants";
 
 export default async function Command(props: LaunchProps<{ arguments: Arguments.Merge }>) {
   try {
@@ -48,6 +47,7 @@ export default async function Command(props: LaunchProps<{ arguments: Arguments.
 
     // Windows implementation using @libpdf/core
     if (isWindows) {
+      const { PDF } = await import("@libpdf/core");
       await closeMainWindow();
 
       await showToast({
