+++ extensions/pdf-tools/src/split-by-page-count.ts
@@ -8,9 +8,8 @@ import {
 } from "@raycast/api";
 import fs from "fs/promises";
 import path from "path";
-import { PDF } from "@libpdf/core";
 import { getSelectedItems } from "universal-selection";
-import { isMac, isWindows, splitByPageCountWindows } from "./lib/utils";
+import { isMac, isWindows } from "./lib/constants";
 
 export default async function Command(props: LaunchProps<{ arguments: Arguments.SplitByPageCount }>) {
   try {
@@ -58,6 +57,9 @@ export default async function Command(props: LaunchProps<{ arguments: Arguments.
 
     // Windows implementation using @libpdf/core
     if (isWindows) {
+      const { PDF } = await import("@libpdf/core");
+      const { splitByPageCountWindows } = await import("./lib/windows");
+
       await closeMainWindow();
 
       for (const item of selectedItems) {
